created activity func

+++ dbo/cartonsetup/cartonsetup.php
@@ -41,36 +41,47 @@ function dbo_cartonsetup_custom_new($table, $cols){
 		else {
 			$newfilename = $newid;	
 			// move file
-			$doc = new Document();
-			$status = $doc->moveSingleImage($filelocation, $newfilename,$image,$newid,'car_id');
-	
-			if (!empty($_POST['detail'])) {
-			
-				foreach($_POST['detail'] as $keycol=>$d) {
-					$currset = false;
-					foreach ($DETAIL_SETUP as $k=>$set) {
-						if (!empty($set['fkcol']) && $set['fkcol'] == $keycol) $currset = $set;
-					}
-					if (empty($d['data']) && $currset['mustfill']) {
-						$ret[] = "Please fill up ".strip_tags($currset['fkcolcaption']);
-					}
-					else {
-						$detcols = array($keycol => $newid);
-						$ok = $DB->execute("delete from {$d['table']} where {$keycol} = :0", array($newid));
-						if(!$ok) $ret[] = $DB->lastError;
+			$except = array("jpg", "jpeg", "gif", "png","JPG", "JPEG", "GIF", "PNG");
+			$imp = implode('|', $except);
+			$imgregexp = '/^.*\.('.$imp.')$/i';
+
+			$setting = array('accept_file_types' => $imgregexp);
+			$doc = new Document($setting);
+
+
+			$fileinfo = $doc->moveSingleImage($filelocation, $newfilename,$image,$newid,'car_id');
+			if(!$fileinfo) $ret[] = $doc->error; 
+			else{
+				$status = $doc->singleDocIns($newid,'car_id',$fileinfo['filename'],$fileinfo['filesize'],$fileinfo['extension']);	
+
+				if (!empty($_POST['detail'])) {
+				
+					foreach($_POST['detail'] as $keycol=>$d) {
+						$currset = false;
+						foreach ($DETAIL_SETUP as $k=>$set) {
+							if (!empty($set['fkcol']) && $set['fkcol'] == $keycol) $currset = $set;
+						}
+						if (empty($d['data']) && $currset['mustfill']) {
+							$ret[] = "Please fill up ".strip_tags($currset['fkcolcaption']);
+						}
 						else {
-							if (!empty($d['data'])) {
-								foreach ($d['data'] as $i=>$currcols) {
-									$detcols = array_merge($detcols, $currcols);
-									foreach ($detcols as &$val) {
-										if (is_array($val)) $val = implode(", ", $val);
-									}
-									foreach ($currset['mustfillCols'] as $mcol) {
-										if (empty($detcols[$mcol])) $ret[] = "Please fill up ".strip_tags($currset['caption'][$mcol]);
-									}
-									if (!$ret) {
-										$ok = $DB->doInsert($d['table'], $detcols);
-										if(!$ok) $ret[] = $DB->lastError;
+							$detcols = array($keycol => $newid);
+							$ok = $DB->execute("delete from {$d['table']} where {$keycol} = :0", array($newid));
+							if(!$ok) $ret[] = $DB->lastError;
+							else {
+								if (!empty($d['data'])) {
+									foreach ($d['data'] as $i=>$currcols) {
+										$detcols = array_merge($detcols, $currcols);
+										foreach ($detcols as &$val) {
+											if (is_array($val)) $val = implode(", ", $val);
+										}
+										foreach ($currset['mustfillCols'] as $mcol) {
+											if (empty($detcols[$mcol])) $ret[] = "Please fill up ".strip_tags($currset['caption'][$mcol]);
+										}
+										if (!$ret) {
+											$ok = $DB->doInsert($d['table'], $detcols);
+											if(!$ok) $ret[] = $DB->lastError;
+										}
 									}
 								}
 							}
